Split sensor.py into three modules (882 → 486 lines)
Extract unit mapping helpers into sensor_helpers.py and diagnostic
entity classes into sensor_diagnostics.py.
